Improving test coverage with new telemetry scopes
