+++ src/tests/Microsoft.Agents.Builder.Tests/Telemetry/TelemetryTestsCollection.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Xunit;
+
+namespace Microsoft.Agents.Builder.Tests.Telemetry
+{
+    /// <summary>
+    /// Disables parallel execution across all telemetry-scope test classes in this assembly
+    /// so that the global <see cref="System.Diagnostics.ActivitySource"/> listener registered
+    /// by each class does not capture activities emitted by another class running concurrently.
+    /// </summary>
+    [CollectionDefinition("TelemetryTests", DisableParallelization = true)]
+    public class TelemetryTestsCollection { }
+}
