+++ src/tests/Microsoft.Agents.Core.Telemetry.Tests/TelemetryTestsCollection.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using Xunit;
+
+namespace Microsoft.Agents.Core.Telemetry.Tests
+{
+    /// <summary>
+    /// Disables parallel execution across all telemetry-scope test classes in this assembly
+    /// so that the global <see cref="System.Diagnostics.ActivitySource"/> listener registered
+    /// by each class does not capture activities emitted by another class running concurrently.
+    /// </summary>
+    [CollectionDefinition("TelemetryScopeTests", DisableParallelization = true)]
+    public class TelemetryTestsCollection { }
+}
